pyproject.toml
[project]
name = "myrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.17",
    "google-genai>=1.30.0",
    "sentence-transformers>=5.1.0",
]
